// src/lib/types.ts
export interface ParsedASSStyle {
    Name: string;
    Fontname: string;
    Fontsize: string;
    PrimaryColour: string;
    SecondaryColour: string;
    OutlineColour: string;
    BackColour: string;
    Bold: string;
    Italic: string;
    Underline: string;
    StrikeOut: string;
    ScaleX: string;
    ScaleY: string;
    Spacing: string;
    Angle: string;
    BorderStyle: string;
    Outline: string;
    Shadow: string;
    Alignment: string;
    MarginL: string;
    MarginR: string;
    MarginV: string;
    Encoding: string;
}